Rename component folder for statement components
