+++ frontend/src/components/Statement/StatementSection.tsx
@@ -1,4 +1,5 @@
 import { Box, Heading } from "@chakra-ui/react";
+import { useColors } from "hooks/useColors";
 import { useLocales } from "hooks/useLocales";
 import { FC, ReactNode } from "react";
 
@@ -9,9 +10,10 @@ interface Props {
 
 const StatementSection: FC<Props> = ({ heading, children }) => {
   const { t } = useLocales();
+  const { boxBorder } = useColors();
 
   return (
-    <Box shadow="sm" p={10} border="1px" borderColor="gray.200" rounded="md">
+    <Box shadow="sm" p={10} border="1px" borderColor={boxBorder} rounded="md">
       <Heading size="md">{heading}</Heading>
       {children}
     </Box>
