+++ frontend/src/components/Statement/StatementSectionTable.tsx
