+++ frontend/src/components/Statement/StatementTableRow.tsx
@@ -12,10 +12,9 @@ import {
   Text,
   Tr
 } from "@chakra-ui/react";
-import { AuthContext } from "contexts/AuthContext";
+import { useColors } from "hooks/useColors";
 import { useLocales } from "hooks/useLocales";
-import { useRouter } from "next/router";
-import { FC, ReactNode, useContext } from "react";
+import { FC } from "react";
 import { BsInfoCircle } from "react-icons/bs";
 
 import InputDKK from "./InputDKK";
@@ -24,14 +23,12 @@ interface Props {
   text: string;
   subText?: string;
   tax: string;
-  children?: ReactNode;
   tooltip?: string;
 }
 
-const StatementTableRow: FC<Props> = ({ text, subText, tax, children, tooltip }) => {
+const StatementTableRow: FC<Props> = ({ text, subText, tax, tooltip }) => {
   const { t } = useLocales();
-  const router = useRouter();
-  const { activeUser } = useContext(AuthContext);
+  const { subTextColor } = useColors();
 
   return (
     <Tr>
@@ -58,7 +55,7 @@ const StatementTableRow: FC<Props> = ({ text, subText, tax, children, tooltip })
           )}
           <Stack spacing={0}>
             <Text>{text}</Text>
-            <Text fontSize="sm" color="gray.500">
+            <Text fontSize="sm" color={subTextColor}>
               {subText}
             </Text>
           </Stack>
@@ -72,26 +69,3 @@ const StatementTableRow: FC<Props> = ({ text, subText, tax, children, tooltip })
   );
 };
 export default StatementTableRow;
-/*
-<Box>
-        {tooltip && (
-          <Popover>
-            <PopoverTrigger>
-              <IconButton
-                aria-label="info"
-                icon={<BsInfoCircle size="20px" />}
-                isRound={true}
-                w="0"
-                size="sm"
-                variant="ghost"
-              />
-            </PopoverTrigger>
-            <PopoverContent>
-              <PopoverArrow />
-              <PopoverCloseButton />
-              <PopoverBody>{tooltip}</PopoverBody>
-            </PopoverContent>
-          </Popover>
-        )}
-      </Box>
-*/
